backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-dotenv>=1.0.0
cua-agent
cua-computer
pydantic>=2.5.0
pydantic-settings>=2.0.0
httpx>=0.25.0
certifi
